remove commented and rename file
