+++ ui/src/app/services/data/data.service.ts
@@ -4,7 +4,7 @@ import IConv, { ICreateIndex, IInterleaveStatus } from '../../model/Conv'
 import IRuleContent from 'src/app/model/Rule'
 import { BehaviorSubject, forkJoin, Observable, of } from 'rxjs'
 import { catchError, filter, map, tap } from 'rxjs/operators'
-import IUpdateTable from 'src/app/model/updateTable'
+import IUpdateTable from 'src/app/model/UpdateTable'
 import IDumpConfig from 'src/app/model/DumpConfig'
 import ISessionConfig from '../../model/SessionConfig'
 import { InputType, StorageKeys } from 'src/app/app.constants'
