+++ ui/src/app/services/fetch/fetch.service.ts
@@ -1,8 +1,8 @@
 import { HttpClient, HttpResponse } from '@angular/common/http'
 import { Injectable } from '@angular/core'
 import IDbConfig from 'src/app/model/DbConfig'
-import ISession, { ISaveSessionPayload } from 'src/app/model/Session'
-import IUpdateTable from 'src/app/model/updateTable'
+import ISession, { ISaveSessionPayload } from '../../model/Session'
+import IUpdateTable from '../../model/UpdateTable'
 import IConv, { ICreateIndex, IInterleaveStatus } from '../../model/Conv'
 import IDumpConfig from '../../model/DumpConfig'
 import ISessionConfig from '../../model/SessionConfig'
